Resolved all warnings
Former-commit-id: 4e80581b4af550c71fd153c4a0aa98b6283905b1

+++ adagucserverEC/CImgWarpGeneric.h
@@ -43,7 +43,7 @@ private:
     double legendUpperRange;
     bool hasNodataValue;
     float *dataField;
-    size_t width, height;
+    int width, height;
   };
 
   template <class T> static void drawFunction(int x, int y, T val, void *_settings, void *g) {
